pyproject.toml
[tool.poetry]
name = "quote_bot"
version = "0.1.0"
description = ""
authors = ["Tommy Heng <tommy.seng.heng@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
SQLAlchemy = "^2.0.31"
discord-py = "^2.4.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
pytest = "^7.3.2"
mypy = "^1.6.1"
flake8 = ">=6.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
style = "scripts.style:main"
start = "scripts.start:main"